+++ tests/integration_tests/test_negative_http_connection.py
@@ -269,14 +269,14 @@ class TestHttpConnectionNegative(TestCase):
         with surreal.connect() as connection:
             res = connection.root_info()
             self.assertTrue(res.is_error())
-            self.assertEqual("Not enough permissions to perform this action", res.result)
+            self.assertEqual("IAM error: Not enough permissions to perform this action", res.result)
 
     def test_ns_info_failed(self):
         surreal = Surreal(URL, 'test', 'test', ('user_db', 'user_db'), use_http=True)
         with surreal.connect() as connection:
             res = connection.ns_info()
             self.assertTrue(res.is_error())
-            self.assertEqual("Not enough permissions to perform this action", res.result)
+            self.assertEqual("IAM error: Not enough permissions to perform this action", res.result)
 
 
     # TODO uncomment after bugfix