+++ ReleaseNotes.md
@@ -1,5 +1,14 @@
 ## Release Notes ##
 
+**Version 0.2.0 (compatible with SurrealDB version 1.1.1):**
+ - QL-constructor added, you can now generate queries and use all features of QL
+ - result now has helpers like is_empty, to_dict, id, ids, get
+ - websocket connection works faster
+ - a lot more examples added
+ - updated readme on all updates and features
+ - bug fixes
+
+
 **Version 0.1.8 (compatible with SurrealDB version 1.1.1):**
 
  - grammar and spelling fixes